Templates, functionality achieved
At this point, I'd managed to make a decent page for light side and dark side character quotes.

Also the core functionality of the quote generator was achieved. Through DOM assignment and the fetch function, the quotes can appear in the central field. However the loading is slow, so I will update the readme with a loading animation task.

Now the focus is a loading animation, and using the 'faction' key to choose between light and dark side quotes. Having tested the API, it seems there are 4 factions, which may mean the original developer created factions for a new set of characters as well. But if the light side/dark side functionality works, it will simply be a matter of adding two more website designs.