Mentor Review - Core functionality implemented
This evening I had a Zoom call with my father. He is an IBM programmer and a great mentor for my programming.

His main criticisms with my code was that I was using ids instead of classes and that my code needs to be more generic and dry. My functions need to serve single purposes.

He helped me refactor my functions to not only fix issues, but also to make it cleaner.

I plan to take his advice in mind with future projects.

The quote generator now shows the specific styling for light side, dark side and neutral when refreshed.