JSON File Update and Conclusion
I've come to the end of the Star Wars Quotes project.  It's been a fun exercise in Javascript and CSS.  As I've juggled it with other work from my bootcamp, it's also helped me learn how to juggle projects and find new solutions to problems.

The issues with the Star Wars API have helped simulate real bug checking in some way, where I had to keep an eye out for solutions to a problem beyond my control. The API is inconsistent, sometimes sending data and sometimes not. I could contact the owner and investigate, but I need to move on to building more important projects. Instead, I found a solution during a portfolio building exercise where a lecturer used a JSON file with an associative array to automate building 'cards', elements in front end that have the same CSS repetitively. His object was almost identical in functionality to the quotes object, so I simply took his design, copied in the associative array structure from the original design, and searched online for new quotes.

It's been fun. On to the next project!